+++ src/components/Modals/index.tsx
@@ -2,5 +2,6 @@ import { AlgorithmsModal } from './AlgorithmsModal';
 import { HelpModal } from './HelpModal';
 import { ImportExportModal } from './ImportExportModal';
 import { Modal } from './Modal';
+import { SettingsModal } from './SettingsModal';
 
-export { AlgorithmsModal, HelpModal, ImportExportModal, Modal };
+export { AlgorithmsModal, HelpModal, ImportExportModal, Modal, SettingsModal };