+++ src/providers/PropertiesProvider.tsx
@@ -5,23 +5,20 @@ import { GraphRegistryProvider } from './GraphRegistryProvider';
 import { GraphSelectionProvider } from './GraphSelectionProvider';
 import { LayoutProvider } from './LayoutProvider';
 import { NodesProvider } from './NodesProvider';
-import { SettingsProvider } from './SettingsProvider';
 
 export function PropertiesProvider({ children }: PropertiesProviderProps) {
     return (
-        <SettingsProvider>
-            <GraphRegistryProvider>
-                <GraphMetaProvider>
-                    <GraphSelectionProvider>
-                        <LayoutProvider>
-                            <NodesProvider>
-                                <EdgesProvider>{children}</EdgesProvider>
-                            </NodesProvider>
-                        </LayoutProvider>
-                    </GraphSelectionProvider>
-                </GraphMetaProvider>
-            </GraphRegistryProvider>
-        </SettingsProvider>
+        <GraphRegistryProvider>
+            <GraphMetaProvider>
+                <GraphSelectionProvider>
+                    <LayoutProvider>
+                        <NodesProvider>
+                            <EdgesProvider>{children}</EdgesProvider>
+                        </NodesProvider>
+                    </LayoutProvider>
+                </GraphSelectionProvider>
+            </GraphMetaProvider>
+        </GraphRegistryProvider>
     );
 }